Move code for the storage initialization
Move code for the storage initialization to
pyanaconda.storage.inialization.
